scape-rag-system/requirements.txt
langchain
langchain-community
langchain-openai
chromadb
tiktoken
python-dotenv
pypdf
streamlit